Reorganization of some functions + Move functions into separate .cpp files + Create corresponding .h files with guards
