+++ src/error.h
@@ -1,0 +1,6 @@
+#ifndef ERROR_H
+#define ERROR_H
+
+void error(std::string s);
+
+#endif
