+++ src/invlogit.h
@@ -1,0 +1,8 @@
+#ifndef INVLOGIT_H
+#define INVLOGIT_H
+
+#include <RcppArmadillo.h>
+
+double invlogit(double x);
+
+#endif
