+++ src/table_cpp.cpp
@@ -1,0 +1,20 @@
+#include <RcppArmadillo.h>
+
+// [[Rcpp::depends(RcppArmadillo)]]
+
+//' Get the top number of observations
+//'
+//' @param v A row vector of values to be tabulated.
+//'
+//' @return A map of (values, frequencies)
+std::map<uint32_t, uint32_t> table_cpp(const arma::urowvec& v) {
+
+  // Create a map to store frequencies
+  std::map<uint32_t, uint32_t> Elt;
+  Elt.clear();
+
+  // Fill the map with occurrences per number.
+  for (int i = 0; i < v.size(); ++i) Elt[ v(i) ] += 1;
+
+  return Elt;
+}
