+++ src/table_cpp.h
@@ -1,0 +1,8 @@
+#ifndef TABLE_CPP_H
+#define TABLE_CPP_H
+
+#include <RcppArmadillo.h>
+
+std::map<uint32_t, uint32_t> table_cpp(const arma::urowvec& v);
+
+#endif
